refactor: extract input-container core CSS to styles folder
