+++ packages/input-container/src/styles/vaadin-input-container-base-styles.js
