+++ packages/input-container/src/styles/vaadin-input-container-core-styles.js
