+++ packages/input-container/src/vaadin-input-container.js
@@ -9,7 +9,7 @@ import { DirMixin } from '@vaadin/component-base/src/dir-mixin.js';
 import { PolylitMixin } from '@vaadin/component-base/src/polylit-mixin.js';
 import { CSSInjectionMixin } from '@vaadin/vaadin-themable-mixin/css-injection-mixin.js';
 import { ThemableMixin } from '@vaadin/vaadin-themable-mixin/vaadin-themable-mixin.js';
-import { inputContainerStyles } from './vaadin-input-container-core-styles.js';
+import { inputContainerStyles } from './styles/vaadin-input-container-core-styles.js';
 
 /**
  * @customElement
